+++ packages/core/src/modules/auth/api/users/route.ts
@@ -38,8 +38,8 @@ const logger = createLogger('auth').child({ component: 'users' })
 
 const querySchema = z.object({
   id: z.string().uuid().optional(),
-  page: z.coerce.number().min(1).default(1),
-  pageSize: z.coerce.number().min(1).max(100).default(50),
+  page: z.coerce.number().int().min(1).default(1),
+  pageSize: z.coerce.number().int().min(1).max(100).default(50),
   search: z.string().optional(),
   name: z.string().optional(),
   organizationId: z.string().uuid().optional(),
