+++ packages/core/src/modules/catalog/api/products/route.ts
@@ -59,8 +59,8 @@ const UUID_REGEX =
 
 const listSchema = z
   .object({
-    page: z.coerce.number().min(1).default(1),
-    pageSize: z.coerce.number().min(1).max(100).default(50),
+    page: z.coerce.number().int().min(1).default(1),
+    pageSize: z.coerce.number().int().min(1).max(100).default(50),
     id: z.string().uuid().optional(),
     search: z.string().optional(),
     status: z.string().optional(),
