+++ packages/core/src/modules/customers/api/assignable-staff/route.ts
@@ -5,8 +5,8 @@ import { createPagedListResponseSchema } from '../openapi'
 
 const querySchema = z
   .object({
-    page: z.coerce.number().min(1).default(1),
-    pageSize: z.coerce.number().min(1).max(100).default(24),
+    page: z.coerce.number().int().min(1).default(1),
+    pageSize: z.coerce.number().int().min(1).max(100).default(24),
     search: z.string().optional(),
   })
   .passthrough()
