+++ packages/core/src/modules/resources/api/areas.ts
@@ -19,8 +19,8 @@ const createInputSchema = resourcesResourceAreaCreateSchema
 
 const listSchema = z
   .object({
-    page: z.coerce.number().min(1).default(1),
-    pageSize: z.coerce.number().min(1).max(100).default(100),
+    page: z.coerce.number().int().min(1).default(1),
+    pageSize: z.coerce.number().int().min(1).max(100).default(100),
     search: z.string().optional(),
     sortField: z.string().optional(),
     sortDir: z.enum(['asc', 'desc']).optional(),
@@ -110,8 +110,8 @@ const crud = makeCrudRoute({
 const viewSchema = z
   .object({
     view: z.enum(['manage', 'tree']).default('manage'),
-    page: z.coerce.number().min(1).default(1),
-    pageSize: z.coerce.number().min(1).max(100).default(100),
+    page: z.coerce.number().int().min(1).default(1),
+    pageSize: z.coerce.number().int().min(1).max(100).default(100),
     search: z.string().optional(),
     status: z.enum(['all', 'active', 'inactive']).optional(),
     areaTypeId: z.string().optional(),
